Well, I think I got it. The type of storage changed to the regular filesystem from the database. Now it's gonna be faster and probably safer? Anyway that was a bad idea to store files in the database unless I need something like changing the file on the fly or creating special HTTP headers... And even then I can simply read the file from the filesystem. So let's check what we've got here. I gonna test.

+++ resources/views/partials/treepage.blade.php
@@ -17,16 +17,18 @@
         <a href="{{ $page->url }}">{{ $page->url }}</a>
     </div>
     <div class="item--controls">
-        {!! Form::open(['action' => ['PagesController@move', $page->id, 'up'], 'method' => 'delete']) !!}
-        <button type="submit" class="btn btn-xs btn-link">
-            <i class="glyphicon glyphicon-arrow-up"></i>
-        </button>
-        {!! Form::close() !!}
-        {!! Form::open(['action' => ['PagesController@move', $page->id,  'down'], 'method' => 'delete']) !!}
-        <button type="submit" class="btn btn-xs btn-link">
-            <i class="glyphicon glyphicon-arrow-down"></i>
-        </button>
-        {!! Form::close() !!}
+        @if ($page->real_level != 0)
+            {!! Form::open(['action' => ['PagesController@move', $page->id, 'up'], 'method' => 'delete']) !!}
+            <button type="submit" class="btn btn-xs btn-link">
+                <i class="glyphicon glyphicon-arrow-up"></i>
+            </button>
+            {!! Form::close() !!}
+            {!! Form::open(['action' => ['PagesController@move', $page->id,  'down'], 'method' => 'delete']) !!}
+            <button type="submit" class="btn btn-xs btn-link">
+                <i class="glyphicon glyphicon-arrow-down"></i>
+            </button>
+            {!! Form::close() !!}
+        @endif
         <a href="/admin/pages/{{ $page->id }}/edit" class="btn btn-xs btn-link">
             <i class="glyphicon glyphicon-edit"></i>
         </a>
